+++ Java25-Team1-FinalProject/src/main/java/com/example/team_project/controllers/TransactionController.java
@@ -3,6 +3,7 @@ package com.example.team_project.controllers;
 import com.example.team_project.entities.Transaction;
 import com.example.team_project.services.TransactionService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,12 +17,13 @@ public class TransactionController {
     @Autowired
     private TransactionService transactionService;
 
+
     // Create a new transaction
     @PostMapping("/create")
-    public ResponseEntity<Transaction> createTransaction(@RequestBody Transaction transaction) {
-        Transaction createdTransaction = transactionService.createNewTransaction(transaction);
-        return ResponseEntity.ok(createdTransaction);
+    public Transaction createTransaction(@RequestBody Transaction transaction) {
+        return transactionService.createNewTransaction(transaction);
     }
+
 
     // Get a list of all transactions
     @GetMapping("/list")
@@ -36,29 +38,30 @@ public class TransactionController {
         if (transaction.isPresent()) {
             return ResponseEntity.ok(transaction.get());
         } else {
-            return ResponseEntity.notFound().build();
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     // Update a transaction by ID
     @PutMapping("update//{id}")
-    public ResponseEntity<Transaction> updateTransaction(@PathVariable Integer id, @RequestBody Transaction transactionDetails) {
-        try {
-            Transaction updatedTransaction = transactionService.updateTransaction(id, transactionDetails);
-            return ResponseEntity.ok(updatedTransaction);
-        } catch (IllegalStateException e) {
-            return ResponseEntity.notFound().build();
+    public ResponseEntity<Transaction> updateTransaction(@PathVariable Integer id, @RequestBody Transaction updatedTransaction) {
+        Optional<Transaction> currentTransaction = transactionService.updateTransaction(id, updatedTransaction);
+        if(currentTransaction.isPresent()){
+            return ResponseEntity.ok(currentTransaction.get());
+        }else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 
     // Delete a transaction by ID
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<Void> deleteTransaction(@PathVariable Integer id) {
-        try {
+        Optional<Transaction> transaction = transactionService.getTransactionById(id);
+        if(transaction.isPresent()){
             transactionService.deleteTransaction(id);
             return ResponseEntity.noContent().build();
-        } catch (IllegalStateException e) {
-            return ResponseEntity.notFound().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
     }
 }
